feat: made use of usecontext for passing down data in the modals :sparkles:
