+++ frontend/src/hooks/modalContext.tsx
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+
+
+
+export const  PromptModalDetsContext  =createContext<any>({});
+export const  AddPromptModalContext= createContext<any>({});
+export const  AddPromptTimesContext = createContext<any>({});
