+++ frontend/src/pages/quests.tsx
@@ -1,6 +1,6 @@
 import { GetServerSideProps, NextPage } from "next";
 import { getServerSession } from "next-auth";
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import { authOptions } from "./api/auth/[...nextauth]";
 import { DashboardLayout, Meta } from "~/components/layouts";
 import { Button, Dialog, DialogContent, DialogDescription, DialogTitle, Input } from "~/components/ui";
@@ -8,7 +8,10 @@ import { api } from "~/config";
 import { useSession } from "next-auth/react";
 import Link from "next/link";
 import category from "~/config/category";
-
+import  {PromptModalDetsContext,AddPromptModalContext,AddPromptTimesContext}  from '~/hooks/modalContext';
+import PromptModalDets from "~/components/questmodals/promptdets";
+import AddPromptModal from "~/components/questmodals/addprompts";
+import AddPromptTimes from "~/components/questmodals/addtimes";
 interface IQuest {
   title: string;
   description: string;
@@ -32,6 +35,9 @@ const QuestsPage: NextPage = () => {
   const [displayShareModal, setDisplayShareModal] = React.useState<boolean>(false);
   const [savedQuests, setSavedQuests] = React.useState<IQuest[]>([]);
   const [activeQuest, setActiveQuest] = React.useState<IQuest | null>(null);
+
+
+  
 
   const saveQuest = async () => {
     try {
@@ -166,194 +172,39 @@ const QuestsPage: NextPage = () => {
   const [displayAddPromptModal, setDisplayAddPromptModal] = useState(false);
   const [displayPromptModalDets, setDisplayPromptModalDets] = useState(false);
   const [displayAddPromptTimes, setDisplayAddPromptTimes] = useState(false);
-  const [selectedCategory, setSelectedCategory] = useState("");
+
+ 
+  const [selectedCategory, setSelectedCategory] = useState<string>('');
+  const [promptText, setPromptText] = useState<string>('');
+  const [responseType, setResponseType] = useState<string>('');
   const [selectedDays, setSelectedDays] = useState<string[]>([]);
-  const [betweenTime, setBetweenTime] = useState<string>("");
-  const [andTime, setAndTime] = useState<string>("");
+  const [betweenTime, setBetweenTime] = useState<string>('');
+  const [andTime, setAndTime] = useState<string>('');
+  const [frequency, setFrequency] = useState<string>('');
 
-  const handleDayChange = (day: string) => {
-    setSelectedDays((prev) => (prev.includes(day) ? prev.filter((d) => d !== day) : [...prev, day]));
-  };
-
-  const handleChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    setSelectedCategory(event.target.value);
-  };
+  console.log("prompt text", {
+    promptText,
+    selectedCategory,
+    responseType,
+    selectedDays,
+    betweenTime,
+    andTime,
+    frequency,
+  });
   const handleAddPromptModal = () => {
     console.log("add prompt modal");
     setDisplayAddPromptModal(true);
   };
 
-  interface AddPromptModalProps {
-    onSave: () => void;
-    onClose: () => void;
-    onContinue: () => void;
-  }
 
-  interface PromptModalDetsProps {
-    onPrevious: () => void;
-    onClose: () => void;
-    onContinue: () => void;
-  }
+ 
 
-  interface AddPromptTimesProps {
-    onPrevious: () => void;
-    onClose: () => void;
-    onContinue: () => void;
-  }
-
-  const AddPromptModal: React.FC<AddPromptModalProps> = ({ onSave, onClose, onContinue }) => {
-    return (
-      <Dialog open={true} onOpenChange={onClose}>
-        <DialogContent>
-          <DialogTitle>Select a category</DialogTitle>
-          <DialogDescription>
-            <div className="list-disc mt-2">
-              <label htmlFor="activity" className="my-2 block text-sm font-medium text-gray-900 dark:text-white mt-4">
-                Choose a category that best describes your prompt.
-              </label>
-              <select value={selectedCategory} onChange={handleChange} className="border rounded-md p-2 my-4 w-full">
-                <option value="" disabled>
-                  Select a category
-                </option>
-                {category.map((item, index) => (
-                  <option key={index} value={item}>
-                    {item}
-                  </option>
-                ))}
-              </select>
-            </div>
-          </DialogDescription>
-          <div className="mt-4 flex justify-end gap-4">
-            <div></div>
-            <Button onClick={onContinue}>Continue</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-    );
-  };
-
-  const PromptModalDets: React.FC<PromptModalDetsProps> = ({ onPrevious, onClose, onContinue }) => {
-    return (
-      <Dialog open={true} onOpenChange={onClose}>
-        <DialogContent>
-          <DialogTitle>Enter Prompt Details</DialogTitle>
-          <DialogDescription>
-            {/* Add content for the details of the prompt */}
-            <div className="list-disc  mt-2">
-              <Input
-                label="Prompt Text"
-                type="text"
-                placeholder="eg : are you feeling energetic"
-                size="md"
-                fullWidth
-                className="mt-4"
-              />
-              <label htmlFor="activity" className="my-2 block text-sm font-medium text-gray-900 dark:text-white mt-4">
-                Select response type:
-                <select
-                  id="activity"
-                  className="block w-full mt-4 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-indigo-500 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-indigo-500 dark:focus:ring-indigo-500"
-                >
-                  <option value="" disabled>
-                    Select a response type
-                  </option>
-                  <option>Yes/No</option>
-                  <option>Text</option>
-                  <option>Number</option>
-                  <option>Custom Option</option>
-                </select>
-              </label>
-            </div>
-          </DialogDescription>
-          <div className="mt-4 flex justify-end space-x-2">
-            <Button onClick={onPrevious}>Previous</Button>
-            <Button onClick={onContinue}>Continue</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-    );
-  };
-
-  const AddPromptTimes: React.FC<AddPromptTimesProps> = ({ onPrevious, onClose, onContinue }) => {
-    return (
-      <Dialog open={true} onOpenChange={onClose}>
-        <DialogContent>
-          <DialogTitle>Set time and frequency</DialogTitle>
-          <DialogDescription>
-            <div className="mt-2">
-              <label htmlFor="frequency" className="block text-sm font-medium text-gray-900 dark:text-white mt-4">
-                How often should we prompt you?
-                <select
-                  id="frequency"
-                  className="block w-full mt-2 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-indigo-500 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-indigo-500 dark:focus:ring-indigo-500"
-                >
-                  <option value="" disabled>
-                    Select an option
-                  </option>
-                  <option>Once</option>
-                  <option>Every 30 minutes</option>
-                  <option>Every hour</option>
-                  <option>Every two hours</option>
-                </select>
-              </label>
-
-              <label htmlFor="time" className="block text-sm font-medium text-gray-900 dark:text-white mt-4">
-                When do you want to be prompted?
-              </label>
-              <div className="mt-2">
-                <label>Between</label>
-                <input
-                  placeholder="HH:MM"
-                  type="time"
-                  id="time"
-                  value={betweenTime}
-                  onChange={(e) => setBetweenTime(e.target.value)}
-                  className="block w-full mt-2 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-indigo-500 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-indigo-500 dark:focus:ring-indigo-500"
-                />
-              </div>
-              <div className="mt-2">
-                <label>And</label>
-                <input
-                  placeholder="5:00 pM"
-                  type="time"
-                  id="time"
-                  value={andTime}
-                  onChange={(e) => setAndTime(e.target.value)}
-                  className="block w-full mt-2 rounded-lg border border-gray-300 bg-gray-50 p-2.5 text-sm text-gray-900 focus:border-indigo-500 focus:ring-indigo-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-indigo-500 dark:focus:ring-indigo-500"
-                />
-              </div>
-              <label className="block text-sm font-medium text-gray-900 dark:text-white mt-6">
-                Which days do you want to be prompted?
-                <div className="flex flex-wrap mt-4 gap-2">
-                  {["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"].map((day) => (
-                    <div key={day} className="flex items-center">
-                      <input
-                        type="checkbox"
-                        id={day}
-                        value={day}
-                        checked={selectedDays.includes(day)}
-                        onChange={() => handleDayChange(day)}
-                        className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
-                      />
-                      <label htmlFor={day} className="ml-2 text-sm font-medium text-gray-900 dark:text-white">
-                        {day}
-                      </label>
-                    </div>
-                  ))}
-                </div>
-              </label>
-            </div>
-          </DialogDescription>
-          <div className="mt-4 flex justify-end space-x-2">
-            <Button onClick={onPrevious}>Previous</Button>
-            <Button onClick={onContinue}>Continue</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-    );
-  };
+ 
 
   return (
+    <AddPromptModalContext.Provider value={{ selectedCategory, setSelectedCategory }}>
+      <AddPromptTimesContext.Provider value ={{ frequency, setFrequency, selectedDays, setSelectedDays, betweenTime, setBetweenTime, andTime, setAndTime }}>
+     < PromptModalDetsContext.Provider value ={{promptText, setPromptText, responseType, setResponseType}}>
     <DashboardLayout>
       <Meta
         meta={{
@@ -526,6 +377,7 @@ const QuestsPage: NextPage = () => {
       )}
       {displayPromptModalDets && (
         <PromptModalDets
+        
           onPrevious={() => {
             setDisplayPromptModalDets(false);
             setDisplayAddPromptModal(true);
@@ -556,6 +408,10 @@ const QuestsPage: NextPage = () => {
         />
       )}
     </DashboardLayout>
+    </PromptModalDetsContext.Provider>
+    </AddPromptTimesContext.Provider>
+    </AddPromptModalContext.Provider>
+  
   );
 };
